# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fabrashion API
  version: 0.1.0
  description: >-
    Backend for **Fabrashion**, a premium clothing store: catalogue, cart,
    orders, and at-home trials.

    This spec is generated from the Zod schemas the server validates against, so
    it always matches the running code.


    ## Authentication

    Everything except `/health` and the auth entry points expects a bearer
    token:


    1. `POST /api/v1/auth/otp/request` with a phone number, then `POST
    /api/v1/auth/otp/verify` with the code → returns `{ user, accessToken,
    refreshToken }`. Phone-OTP is the only login method; first verify creates
    the account.

    2. Send `Authorization: Bearer <accessToken>` on subsequent calls.

    3. The access token lasts **15 minutes**. On a 401, call `POST
    /api/v1/auth/refresh` with your refresh token and retry.


    Refresh tokens are **single-use and rotated**: each refresh returns a new
    one and kills the old. Persist the new pair every time, and never run two
    refreshes concurrently — replaying a rotated token is read as theft and
    signs the user out of every device.


    ## Errors

    Every failure uses one envelope — see the `Error` schema:


    ```json

    { "error": { "code": "UNAUTHORIZED", "message": "Invalid email or password"
    } }

    ```


    Branch on `error.code` (stable), never on `message` (copy, may change). A
    422 adds `details.fieldErrors`, mapping each rejected field to its messages
...


...

...

...

...

...

...


...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Short-lived (15 min) access token from `POST /auth/otp/verify`,
        sent as `Authorization: Bearer <accessToken>`. When it expires the API
        returns 401 UNAUTHORIZED — exchange the refresh token at `POST
        /auth/refresh` for a new pair."
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - VALIDATION_ERROR
                - TOO_MANY_REQUESTS
                - INTERNAL_ERROR
              description: Stable, machine-readable code. Branch on this, never on `message`.
              example: UNAUTHORIZED
            message:
              type: string
              description: Human-readable explanation. Safe to surface to the user.
              example: Invalid email or password
            details:
              type: object
              description: Optional, code-specific context. See `ValidationError` for the 422
                shape.
          required:
            - code
            - message
      required:
        - error
      description: Every non-2xx response from this API uses this envelope.
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: object
              properties:
                formErrors:
                  type: array
                  items:
                    type: string
                  description: Errors not attributable to a single field.
                  example: []
                fieldErrors:
                  type: object
                  additionalProperties:
                    type: array
                    items:
                      type: string
                  description: Field name → list of messages. Render these next to the input.
                  example:
                    email:
                      - Enter a valid email
                    password:
                      - Password must be at least 8 characters
              required:
                - formErrors
                - fieldErrors
          required:
            - code
            - message
            - details
      required:
        - error
      description: Request body failed schema validation.
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: b1d1f0a4-4c9e-4b9a-9c1e-2f3a4b5c6d7e
        fullName:
          type: string
          example: Aarav Sharma
        email:
          type: string
          nullable: true
          format: email
          description: Null for phone-only accounts (Phase 1b).
          example: aarav.sharma@example.com
        phone:
          type: string
          nullable: true
          description: E.164, always +91-prefixed. Null for email-only accounts.
          example: "+919876543210"
        role:
          type: string
          enum:
            - CUSTOMER
            - STAFF
            - ADMIN
          description: Authorisation role. Customers may only touch their own resources.
          example: CUSTOMER
      required:
        - id
        - fullName
        - email
        - phone
        - role
      description: The authenticated user. Never contains credentials.
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
          description: "JWT, valid 15 minutes. Send as `Authorization: Bearer
            <accessToken>`."
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJiMWQxZjBhNC00YzllLTRiOWEtOWMxZS0yZjNhNGI1YzZkN2UiLCJyb2xlIjoiQ1VTVE9NRVIiLCJpYXQiOjE3NTIzOTA0MDAsImV4cCI6MTc1MjM5MTMwMH0.9x2QpQ0rQ6m2mVQ0F6bqK3nqfQ0X1yQ0m8rQ0y9kQ2s
        refreshToken:
          type: string
          description: "Opaque token, valid 30 days, single-use: every refresh rotates it.
            Store it in secure storage (Keychain / Keystore), never in plain
            AsyncStorage — and replace it with the one each refresh returns."
          example: 3f6d1c9a5b8e4d2f7a0c1e9b4d6f8a2c5e7b9d1f3a5c7e9b1d3f5a7c9e1b3d5f
      required:
        - user
        - accessToken
        - refreshToken
      description: The signed-in user plus a new token pair.
    OtpRequestResponse:
      type: object
      properties:
        expiresInSec:
          type: integer
          description: Seconds until the code expires. Drives the resend countdown in the
            app.
          example: 300
        devCode:
          type: string
          description: The OTP itself, so the flow is walkable without an SMS provider.
            Only present when `OTP_EXPOSE_CODE` is on and `NODE_ENV` is not
            production — never in production.
          example: "1234"
      required:
        - expiresInSec
      description: Confirmation that an OTP was sent. Never contains the user.
    OtpRequest:
      type: object
      properties:
        phone:
          type: string
          pattern: ^[6-9]\d{9}$
          description: "Indian mobile number: 10 digits starting 6-9, no country code.
            Stored as +91XXXXXXXXXX."
          example: "9876543210"
      required:
        - phone
    OtpVerifyRequest:
      type: object
      properties:
        phone:
          type: string
          pattern: ^[6-9]\d{9}$
          description: "Indian mobile number: 10 digits starting 6-9, no country code.
            Stored as +91XXXXXXXXXX."
          example: "9876543210"
        code:
          type: string
          pattern: ^\d{4}$
          description: The 4-digit code sent by SMS.
          example: "1234"
      required:
        - phone
        - code
...
paths:
...
  /api/v1/auth/otp/request:
    post:
      operationId: requestOtp
      tags:
        - Auth
      summary: Send a login OTP to a phone number
      security: []
      description: >-
        Sends a 4-digit code by SMS to the given mobile number. This is step 1
        of the only login method — there is no password anywhere in the API;
        email is profile data, never a credential.


        The response is deliberately identical whether or not the number already
        has an account: the API will not confirm whether a phone is registered.
        An account is created on first successful verify.


        A resend cooldown applies on top of the route rate limiter, so calling
        this in a tight loop returns 429.
      requestBody:
        required: true
        description: The mobile number to send the code to.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: Code sent (or re-sent). Prompt the user for it, then call
            `/auth/otp/verify`.
          headers: &a7
            x-request-id: *a4
            RateLimit: &a5
              description: Remaining quota for the current window, e.g. `limit=30,
                remaining=27, reset=840`.
              schema:
                type: string
                example: limit=30, remaining=27, reset=840
            RateLimit-Policy: &a6
              description: The policy in force, e.g. `30;w=900` (30 requests per 900s).
              schema:
                type: string
                example: 30;w=900
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResponse"
              example:
                expiresInSec: 300
                devCode: "1234"
        "422":
          description: The request body failed validation. `details.fieldErrors` maps
            field → messages.
          headers: *a2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              example: &a8
                error:
                  code: VALIDATION_ERROR
                  message: Validation failed
                  details:
                    formErrors: []
                    fieldErrors:
                      email:
                        - Enter a valid email
                      phone:
                        - Enter a valid 10-digit mobile number
        "429":
          description: "Rate limit exceeded: 30 requests per 15 minutes per IP across the
            auth endpoints. Back off and retry after `Retry-After` seconds."
          headers: &a9
            x-request-id: *a4
            RateLimit: *a5
            RateLimit-Policy: *a6
            Retry-After:
              description: Seconds to wait before retrying.
              schema:
                type: integer
                example: 840
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: &a10
                error:
                  code: TOO_MANY_REQUESTS
                  message: Too many attempts. Please try again later.
        "500":
          description: Unexpected server error. Nothing is leaked; check logs by
            `x-request-id`.
          headers: *a2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: *a3
  /api/v1/auth/otp/verify:
    post:
      operationId: verifyOtp
      tags:
        - Auth
      summary: Verify the OTP and sign in
      security: []
      description: >-
        Step 2: exchanges a phone + code for a token pair. On the very first
        successful verify for a number, a `CUSTOMER` account is created — so
        this is both sign-up and sign-in.


        Existing `STAFF` / `ADMIN` numbers sign in with their stored role, which
        is what the app uses to pick the shell (customer, staff, or admin).


        Codes are single-use and expire (`OTP_TTL_MINUTES`). Too many wrong
        attempts (`OTP_MAX_ATTEMPTS`) burns the challenge — request a new code.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: Signed in.
          headers: *a7
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
              example:
                user: &a11
                  id: b1d1f0a4-4c9e-4b9a-9c1e-2f3a4b5c6d7e
                  fullName: Aarav Sharma
                  email: aarav.sharma@example.com
                  phone: "+919876543210"
                  role: CUSTOMER
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJiMWQxZjBhNC00YzllLTRiOWEtOWMxZS0yZjNhNGI1YzZkN2UiLCJyb2xlIjoiQ1VTVE9NRVIiLCJpYXQiOjE3NTIzOTA0MDAsImV4cCI6MTc1MjM5MTMwMH0.9x2QpQ0rQ6m2mVQ0F6bqK3nqfQ0X1yQ0m8rQ0y9kQ2s
                refreshToken: 3f6d1c9a5b8e4d2f7a0c1e9b4d6f8a2c5e7b9d1f3a5c7e9b1d3f5a7c9e1b3d5f
        "401":
          description: The code is wrong, expired, already used, or out of attempts.
          headers: *a2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: UNAUTHORIZED
                  message: Invalid or expired code
        "422":
          description: The request body failed validation. `details.fieldErrors` maps
            field → messages.
          headers: *a2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
              example: *a8
        "429":
          description: "Rate limit exceeded: 30 requests per 15 minutes per IP across the
            auth endpoints. Back off and retry after `Retry-After` seconds."
          headers: *a9
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: *a10
        "500":
          description: Unexpected server error. Nothing is leaked; check logs by
            `x-request-id`.
          headers: *a2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example: *a3
...